@@ -2,54 +2,52 @@
   "name": "horext-ssr-client",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "nuxt",
     "build": "nuxt build",
-    "start": "nuxt start",
+    "dev": "nuxt dev",
     "generate": "nuxt generate",
-    "lint:js": "eslint --ext \".js,.vue\" --ignore-path .gitignore .",
-    "lint": "yarn lint:js",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare",
+    "lint:js": "eslint --ext \".js,.ts,.vue\" --ignore-path .gitignore .",
+    "lint:prettier": "prettier --check --ignore-path .gitignore \"**/*.{js,ts,vue}\"",
+    "lint": "yarn lint:js && yarn lint:prettier",
+    "lint:fix": "prettier --write --ignore-path .gitignore \"**/*.{js,ts,vue}\" && yarn lint:js --fix",
+    "test:unit": "jest --coverage",
     "test": "jest"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.13.6",
-    "@nuxtjs/composition-api": "^0.33.1",
-    "@nuxtjs/universal-storage": "^0.5.9",
-    "@pinia/nuxt": "0.2.1",
-    "@types/luxon": "^2.0.0",
-    "@types/uuid": "^8.3.1",
-    "axios": "^0.24.0",
-    "core-js": "^3.15.1",
+    "@mdi/font": "^7.4.47",
+    "@pinia/nuxt": "0.5.1",
+    "@vueuse/core": "^10.7.1",
+    "@vueuse/nuxt": "^10.7.1",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
     "lottie-web": "^5.10.2",
     "luxon": "^2.0.2",
-    "nuxt": "^2.17.1",
-    "pinia": "^2.1.4",
+    "pinia": "2.1.7",
     "uuid": "^8.3.2",
-    "vue": "2.7.14",
-    "vuetify": "^v2.7.1"
+    "vite": "^5.0.10",
+    "vue": "^3.4.3",
+    "vuetify": "^3.4.9"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.14.7",
-    "@nuxt/types": "^2.17.1",
-    "@nuxt/typescript-build": "^3.0.1",
-    "@nuxtjs/eslint-config-typescript": "^9.0.0",
-    "@nuxtjs/eslint-module": "^3.0.2",
-    "@nuxtjs/google-analytics": "^2.4.0",
-    "@nuxtjs/vuetify": "^1.12.3",
-    "@vue/compiler-sfc": "2.7.14",
-    "@vue/test-utils": "^2.0.0-rc.17",
-    "@vueuse/core": "^10.3.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "^27.0.5",
-    "eslint": "^8.12.0",
-    "eslint-plugin-nuxt": "^3.2.0",
-    "eslint-plugin-vue": "^8.5.0",
-    "jest": "^27.0.5",
-    "sass": "1.63.3",
-    "ts-jest": "^27.0.3",
+    "@nuxtjs/eslint-config-typescript": "^12.1.0",
+    "@types/luxon": "^2.0.0",
+    "@types/uuid": "^8.3.1",
+    "@typescript-eslint/eslint-plugin": "^6.16.0",
+    "@vue/test-utils": "^2.4.3",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-nuxt": "^4.0.0",
+    "eslint-plugin-vue": "^9.19.2",
+    "eslint-plugin-vuetify": "^2.1.1",
+    "nuxt": "^3.9.0",
+    "prettier": "^3.1.1",
+    "sass": "^1.69.6",
     "typescript": "^5.1.6",
-    "vue-jest": "^3.0.4"
+    "vite-plugin-vuetify": "^2.0.1",
+    "vue-tsc": "^1.8.27"
   }
 }
