@@ -1,18 +1,10 @@
 <template>
-  <v-navigation-drawer
-    bottom
-    color="transparent"
-    fixed
-    height="auto"
-    overlay-color="secondary"
-    overlay-opacity=".8"
-    temporary
-    v-bind="$attrs"
-    v-on="$listeners"
-  >
+  <v-navigation-drawer location="bottom" scrim="secondary" temporary>
     <v-list
-      :color="$vuetify.theme.dark?'grey darken-4':'grey lighten-5'"
-      shaped
+      :color="
+        $vuetify.theme.name === 'dark' ? 'grey darken-4' : 'grey lighten-5'
+      "
+      variant="outlined"
     >
       <v-list-item
         v-for="item in items"
@@ -20,25 +12,25 @@
         :to="item.route"
         color="primary"
       >
-        <v-list-item-content>
-          <v-list-item-title v-text="item.name" />
-        </v-list-item-content>
+        <v-list-item-title>
+          {{ item.name }}
+        </v-list-item-title>
       </v-list-item>
     </v-list>
   </v-navigation-drawer>
 </template>
 
 <script lang="ts">
-import Vue from 'vue'
+import { defineComponent } from 'vue'
 
-export default Vue.extend({
+export default defineComponent({
   name: 'HomeDrawer',
 
   props: {
     items: {
       type: Array,
-      default: () => ([])
-    }
-  }
+      default: () => [],
+    },
+  },
 })
 </script>
