@@ -1,26 +1,16 @@
 <template>
-  <v-footer id="home-footer" min-height="48">
+  <v-footer id="home-footer" app>
     <v-container>
       <v-row dense>
         <v-col cols="12" md="6">
           <div
-            class="
-              d-flex
-              flex-wrap
-              justify-md-start justify-center justify-md-none
-            "
+            class="d-flex flex-wrap justify-md-start justify-center justify-md-none"
           >
-            <template v-for="(s, i) in social">
-              <a
-                :key="s.name"
-                class="pa-1 pa-md-0"
-                :href="s.link"
-                v-text="s.name"
-              />
+            <template v-for="(s, i) in social" :key="s.name">
+              <a class="pa-1 pa-md-0" :href="s.link" v-text="s.name" />
 
               <v-responsive
                 v-if="i < social.length - 1"
-                :key="`divider-${s}`"
                 class="mx-4 shrink hidden-sm-and-down"
                 max-height="16"
               >
@@ -32,19 +22,13 @@
         <v-col class="text-center text-md-right" cols="12" md="6">
           <v-row justify-md="end" justify="center" dense>
             <v-col cols="auto">
-              <nuxt-link to="/security">
-                Seguridad
-              </nuxt-link>
+              <nuxt-link to="/security"> Seguridad </nuxt-link>
             </v-col>
             <v-col cols="auto">
-              <nuxt-link to="/terms">
-                Términos y Condiciones
-              </nuxt-link>
+              <nuxt-link to="/terms"> Términos y Condiciones </nuxt-link>
             </v-col>
             <v-col cols="auto">
-              <nuxt-link to="/privacy">
-                Politica de Privacidad
-              </nuxt-link>
+              <nuxt-link to="/privacy"> Politica de Privacidad </nuxt-link>
             </v-col>
           </v-row>
         </v-col>
@@ -57,26 +41,27 @@
 </template>
 
 <script lang="ts">
-import Vue from 'vue'
+import { defineComponent } from 'vue'
 
-export default Vue.extend({
+export default defineComponent({
   name: 'HomeFooter',
   data: () => ({
     social: [
       {
         name: 'Facebook',
-        link: 'https://www.facebook.com/OctatecOficial'
+        link: 'https://www.facebook.com/OctatecOficial',
       },
       {
         name: 'Linkedin',
-        link: 'https://www.linkedin.com/in/franz-antony-bendezu-isidro/'
-      }
-    ]
-  })
+        link: 'https://www.linkedin.com/in/franz-antony-bendezu-isidro/',
+      },
+    ],
+  }),
 })
 </script>
 
-<style lang="sass">
-#home-footer a
-  text-decoration: none
+<style>
+#home-footer a {
+  text-decoration: none;
+}
 </style>
