@@ -1,7 +1,7 @@
 <template>
   <v-sheet
     :color="event.color"
-    class="white--text   text-wrap pa-1 ma-0 "
+    class="text-white text-wrap pa-1 ma-0"
     style="max-width: inherit"
   >
     <template v-if="!event.code">
@@ -15,17 +15,12 @@
         {{ event.section }}
         ({{ event.type }})
       </span>
-      <br>
+      <br />
       {{ event.name }}
     </template>
-    <br>
+    <br />
     <v-fade-transition>
-      <v-overlay
-        v-if="hover"
-        z-index="0"
-        absolute
-        :color="event.color"
-      >
+      <v-overlay v-if="hover" z-index="0" absolute :scrim="event.color">
         <v-icon>mdi-information</v-icon>
       </v-overlay>
     </v-fade-transition>
@@ -36,8 +31,8 @@ export default {
   name: 'ScheduleEventInfo',
   props: {
     event: {},
-    hover: {}
-  }
+    hover: {},
+  },
 }
 </script>
 <style lang="sass">
