@@ -4,7 +4,7 @@
       <v-btn
         :loading="loadingPdf"
         color="success"
-        dark
+        theme="dark"
         class="ma-1"
         @click="downloadPdf"
       >
@@ -17,7 +17,7 @@
       <v-btn
         :loading="loadingImage"
         color="success"
-        dark
+        theme="dark"
         class="ma-1"
         @click="downloadImage"
       >
@@ -32,30 +32,25 @@
 <script lang="ts">
 import { defineComponent, ref } from 'vue'
 import { exportToPNG, exportToPDF } from '~/utils/exportToPNG'
-import { useVuetify } from '~/composables/vuetify'
 
 export default defineComponent({
-  setup () {
-    const vuetify = useVuetify()
-
+  setup() {
     const loading = ref(false)
     const loadingPdf = ref(false)
     const loadingImage = ref(false)
 
-    function getCalendar (): HTMLElement | null {
+    function getCalendar(): HTMLElement | null {
       return document.getElementById('calendar')
     }
 
-    async function downloadImage () {
+    async function downloadImage() {
       loadingImage.value = true
-      await vuetify.goTo(0)
       await exportToPNG(getCalendar())
       loadingImage.value = false
     }
 
-    async function downloadPdf () {
+    async function downloadPdf() {
       loadingPdf.value = true
-      await vuetify.goTo(0)
       await exportToPDF(getCalendar())
       loadingPdf.value = false
     }
@@ -67,12 +62,10 @@ export default defineComponent({
       loadingPdf,
       loadingImage,
       downloadImage,
-      downloadPdf
+      downloadPdf,
     }
-  }
+  },
 })
 </script>
 
-<style scoped>
-
-</style>
+<style scoped></style>
