@@ -1,38 +1,36 @@
 <template>
   <div>
     <v-btn
-      fab
-      outlined
+      variant="outlined"
       icon
-      :color="isFavorite ? 'yellow' : null"
+      :color="isFavorite ? 'yellow' : undefined"
       @click="changeFavoriteState"
     >
-      <v-icon :color="isFavorite ? 'yellow' : null">
-        mdi-star
-      </v-icon>
+      <v-icon :color="isFavorite ? 'yellow' : undefined"> mdi-star </v-icon>
     </v-btn>
   </div>
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, PropType, ref } from 'vue'
+import { computed, defineComponent, type PropType, ref } from 'vue'
 import { useVModel } from '@vueuse/core'
 import { useSnackbar } from '~/composables/snackbar'
+import type { ISchedule } from '~/interfaces/schedule'
 
 export default defineComponent({
   name: 'ScheduleFavoriteAdd',
   props: {
     schedule: {
-      type: Object as PropType<any>,
-      default: null
+      type: Object as PropType<ISchedule>,
+      required: true,
     },
     favoritesSchedules: {
-      type: Array as PropType<Array<any>>,
-      default: () => []
-    }
+      type: Array as PropType<Array<ISchedule>>,
+      default: () => [],
+    },
   },
   emits: ['update:schedule', 'update:favoritesSchedules'],
-  setup (props, { emit }) {
+  setup(props, { emit }) {
     const currentSchedule = useVModel(props, 'schedule', emit)
 
     const favoritesSchedulesSync = useVModel(props, 'favoritesSchedules', emit)
@@ -49,7 +47,7 @@ export default defineComponent({
         } else {
           favoritesSchedulesSync.value = [
             ...favoritesSchedulesSync.value,
-            props.schedule
+            props.schedule,
           ]
         }
         snackbar.showMessage({
@@ -57,7 +55,7 @@ export default defineComponent({
             ? 'Agregado a favoritos'
             : 'Quitado de Favoritos',
           timeout: 2000,
-          color: 'yellow darken-3'
+          color: 'yellow darken-3',
         })
       }
     }
@@ -67,7 +65,7 @@ export default defineComponent({
     const indexSchedule = computed(() => {
       if (currentSchedule.value) {
         return favoritesSchedulesSync.value.findIndex(
-          e => e && e.id === currentSchedule.value.id
+          (e) => e && e.id === currentSchedule.value.id,
         )
       } else {
         return -1
@@ -81,10 +79,8 @@ export default defineComponent({
       message,
       changeFavoriteState,
       isFavorite,
-      indexSchedule
+      indexSchedule,
     }
-  }
+  },
 })
 </script>
-
-<style scoped></style>
