@@ -2,7 +2,7 @@
   <v-snackbar v-model="show" :color="color" :timeout="timeout">
     {{ message }}
     <template #action="{ attrs }">
-      <v-btn text v-bind="attrs" @click="show = false">
+      <v-btn variant="text" v-bind="attrs" @click="show = false">
         Cerrar
       </v-btn>
     </template>
@@ -10,10 +10,10 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from '@nuxtjs/composition-api'
+import { defineComponent } from 'vue'
 import { setupSnackbar } from '~/composables/snackbar'
 export default defineComponent({
   name: 'TheSnackbar',
-  setup: setupSnackbar
+  setup: setupSnackbar,
 })
 </script>
