@@ -1,39 +1,20 @@
 <template>
-  <v-simple-table dense>
+  <v-table dense>
     <template #default>
       <thead>
         <tr>
-          <th class="text-left">
-            Nombre del Curso
-          </th>
-          <th class="text-left">
-            Curso / Sección
-          </th>
-          <th class="text-left">
-            Día
-          </th>
-          <th class="text-left">
-            Horas
-          </th>
-          <th class="text-left">
-            Docente
-          </th>
-          <th class="text-left">
-            Tipo
-          </th>
-          <th class="text-left">
-            Aula
-          </th>
+          <th class="text-left">Nombre del Curso</th>
+          <th class="text-left">Curso / Sección</th>
+          <th class="text-left">Día</th>
+          <th class="text-left">Horas</th>
+          <th class="text-left">Docente</th>
+          <th class="text-left">Tipo</th>
+          <th class="text-left">Aula</th>
         </tr>
       </thead>
       <tbody>
-        <template
-          v-for="item in schedule.schedule"
-        >
-          <schedule-section
-            :key="item.id"
-            :schedule="item"
-          />
+        <template v-for="item in schedule.schedule" :key="item.id">
+          <schedule-section :schedule="item" />
           <class-session-item
             v-for="session in item.sessions"
             :key="session.id"
@@ -42,11 +23,11 @@
         </template>
       </tbody>
     </template>
-  </v-simple-table>
+  </v-table>
 </template>
 
 <script lang="ts">
-import { defineComponent } from '@nuxtjs/composition-api'
+import { defineComponent } from 'vue'
 import ScheduleSection from './ViewListSubjectSection.vue'
 import ClassSessionItem from '~/components/subject/ClassSessionItem.vue'
 
@@ -56,8 +37,8 @@ export default defineComponent({
   props: {
     schedule: {
       type: Object,
-      required: true
-    }
-  }
+      required: true,
+    },
+  },
 })
 </script>
