@@ -1,11 +1,7 @@
 <template>
   <tbody>
-    <template v-for="schedule in schedules">
-      <ScheduleSection
-        :key="schedule.id"
-        v-model="valueSync"
-        :schedule="schedule"
-      />
+    <template v-for="schedule in schedules" :key="schedule.id">
+      <ScheduleSection v-model="valueSync" :schedule="schedule" />
       <ClassSessionItem
         v-for="session in schedule.sessions"
         :key="session.id"
@@ -16,33 +12,32 @@
 </template>
 
 <script lang="ts">
-import { PropType } from '@nuxtjs/composition-api'
+import type { PropType } from 'vue'
 import { useVModel } from '@vueuse/core'
 import { defineComponent } from 'vue'
 import ClassSessionItem from '~/components/subject/ClassSessionItem.vue'
 import ScheduleSection from '~/components/subject/ScheduleSection.vue'
+import type { ISubjectSchedule } from '~/interfaces/subject'
 
 export default defineComponent({
   components: { ScheduleSection, ClassSessionItem },
   props: {
     schedules: {
-      type: Array as PropType<any[]>,
-      required: true
+      type: Array as PropType<ISubjectSchedule[]>,
+      required: true,
     },
-    value: {
-      type: [Array, Object],
-      required: true
-    }
+    modelValue: {
+      type: Array as PropType<ISubjectSchedule[]>,
+      required: true,
+    },
   },
-  emits: ['input'],
-  setup (props, { emit }) {
-    const valueSync = useVModel(props, 'value', emit, {
-      eventName: 'input'
-    })
+  emits: ['update:modelValue'],
+  setup(props, { emit }) {
+    const valueSync = useVModel(props, 'modelValue', emit)
 
     return {
-      valueSync
+      valueSync,
     }
-  }
+  },
 })
 </script>
