@@ -1,50 +1,31 @@
 <template>
-  <v-card :loading="$fetchState.pending">
+  <v-card :loading="pending">
     <v-card-title>
-      <span class="headline">{{ title }}</span>
+      <span class="text-h5">{{ title }}</span>
     </v-card-title>
     <v-card-text>
-      <v-simple-table dense>
+      <v-table dense>
         <template #default>
           <thead>
             <tr>
-              <th class="text-left">
-                Sección
-              </th>
-              <th class="text-left">
-                Día
-              </th>
-              <th class="text-left">
-                Horas
-              </th>
-              <th class="text-left">
-                Docente
-              </th>
-              <th class="text-left">
-                Tipo
-              </th>
-              <th class="text-left">
-                Aula
-              </th>
+              <th class="text-left">Sección</th>
+              <th class="text-left">Día</th>
+              <th class="text-left">Horas</th>
+              <th class="text-left">Docente</th>
+              <th class="text-left">Tipo</th>
+              <th class="text-left">Aula</th>
             </tr>
           </thead>
-          <ScheduleSubjectList
-            v-model="selected"
-            :schedules="schedules"
-          />
+          <ScheduleSubjectList v-model="selected" :schedules="schedules" />
         </template>
-      </v-simple-table>
+      </v-table>
     </v-card-text>
     <v-card-actions>
       <v-spacer />
-      <v-btn
-        color="primary"
-        text
-        @click="$emit('cancel')"
-      >
+      <v-btn color="primary" variant="text" @click="$emit('cancel')">
         Cancelar
       </v-btn>
-      <v-btn color="primary" text @click="saveSections">
+      <v-btn color="primary" variant="text" @click="saveSections">
         Guardar
       </v-btn>
     </v-card-actions>
@@ -52,53 +33,77 @@
 </template>
 
 <script lang="ts">
-import { useFetch } from '@nuxtjs/composition-api'
 import { computed, defineComponent, onMounted, ref, watch } from 'vue'
+import type { PropType } from 'vue'
 import ScheduleSubjectList from '~/components/subject/ScheduleItem.vue'
-import { $api } from '~/utils/api'
+import { useApi } from '~/composables/api'
+import type { IHourlyLoad } from '~/interfaces/houly-load'
+import type { IScheduleSubject } from '~/interfaces/schedule-subject'
+import type {
+  ISelectedSubject,
+  ISession,
+  ISubjectSchedule,
+} from '~/interfaces/subject'
 
 export default defineComponent({
   components: { ScheduleSubjectList },
   props: {
-    subject: { type: Object, required: true },
-    hourlyLoad: { type: Object, required: true }
+    subject: {
+      type: Object as PropType<ISelectedSubject>,
+      required: true,
+    },
+    hourlyLoad: {
+      type: Object as PropType<IHourlyLoad>,
+      required: true,
+    },
   },
-  setup (props, { emit }) {
-    const selected = ref<any[]>([])
-    const schedulesSubject = ref<any[]>([])
-    const sessions = ref<any[]>([])
+  setup(props, { emit }) {
+    const $api = useApi()
+    const selected = ref<ISubjectSchedule[]>([])
+    const schedulesSubject = ref<IScheduleSubject[]>([])
+    const sessions = ref<ISession[]>([])
 
     const fetchSchedules = async () => {
       if (props.subject?.id && props.hourlyLoad?.id) {
-        const { data: schedulesSubjectData } = await $api.scheduleSubject.findBySubjectIdAndHourlyLoadId(
-          props.subject.id, props.hourlyLoad.id
-        )
-        const ids = schedulesSubjectData.map((sb:any) => sb.schedule.id)
-        const { data: sessionsData } = await $api.classSessions.findScheduleIds(ids)
+        const schedulesSubjectData =
+          await $api.scheduleSubject.findBySubjectIdAndHourlyLoadId(
+            props.subject.id,
+            props.hourlyLoad.id,
+          )
+        const ids = schedulesSubjectData.map((sb) => sb.schedule.id)
+        const sessionsData = await $api.classSessions.findScheduleIds(ids)
         sessions.value = sessionsData
         schedulesSubject.value = schedulesSubjectData
       }
     }
-    const { fetch } = useFetch(fetchSchedules)
+    const { refresh, pending } = useAsyncData(fetchSchedules)
 
-    const schedules = computed(() => {
-      return schedulesSubject.value.map(sb => ({
+    const schedules = computed<ISubjectSchedule[]>(() => {
+      return schedulesSubject.value.map((sb) => ({
         ...sb?.schedule,
         scheduleSubject: {
-          id: sb.id
+          id: sb.id,
         },
-        sessions: sessions.value.filter(s => s.schedule.id === sb.schedule.id)
+        sessions: sessions.value.filter(
+          (s) => s.schedule.id === sb.schedule.id,
+        ),
+        subject: props.subject,
       }))
     })
 
-    watch(() => props.subject, () => {
-      fetch()
-    })
+    watch(
+      () => props.subject,
+      () => {
+        refresh()
+      },
+    )
 
     watch(schedules, () => {
       if (props.subject.schedules) {
         selected.value = schedules.value.filter((s1) => {
-          const schedule = props.subject.schedules.find((s2: any) => s2.section.id === s1.section.id)
+          const schedule = props.subject.schedules.find(
+            (s2) => s2.section.id === s1.section.id,
+          )
           return schedule?.id === s1?.id
         })
       } else {
@@ -122,20 +127,9 @@ export default defineComponent({
       selected,
       schedules,
       saveSections,
-      title
+      title,
+      pending,
     }
-  }
+  },
 })
 </script>
-
-<style lang="sass">
-@import '~vuetify/src/styles/styles.sass'
-@media #{map-get($display-breakpoints, 'sm-and-down')}
-  .v-data-table
-    td, th
-      padding: 0 10px
-
-    td
-      font-size: 0.75rem
-
-</style>
