@@ -1,106 +1,75 @@
-import type { NuxtConfig } from '@nuxt/types'
+import vuetify from 'vite-plugin-vuetify'
 
-const config: NuxtConfig = {
+const config = defineNuxtConfig({
   // Global page headers: https://go.nuxtjs.dev/config-head
-  head: {
-    titleTemplate: '%s - Horext',
-    title: 'Bienvenido',
-    meta: [
-      { charset: 'utf-8' },
-      { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      {
-        hid: 'description',
-        name: 'description',
-        content:
-          'Horext es una aplicación web que te ayuda a generar horarios '
-      },
-      { name: 'format-detection', content: 'telephone=no' }
-    ],
-    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }]
+  app: {
+    head: {
+      titleTemplate: '%s - Horext',
+      title: 'Bienvenido',
+      meta: [
+        { charset: 'utf-8' },
+        { name: 'viewport', content: 'width=device-width, initial-scale=1' },
+        {
+          hid: 'description',
+          name: 'description',
+          content:
+            'Horext es una aplicación web que te ayuda a generar horarios ',
+        },
+        { name: 'format-detection', content: 'telephone=no' },
+      ],
+      link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }],
+    },
   },
-
   // Global CSS: https://go.nuxtjs.dev/config-css
-  css: [],
-
-  // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
-  plugins: [
-    '~/plugins/api.ts',
-    '~/plugins/snackbar.ts',
-    '~/plugins/snackbar-accessor.ts',
-    { src: '~/plugins/html2canvas.client.js', mode: 'client' },
-    '~/plugins/axios-accesor',
-    '~/plugins/storage-accessor',
-    '~/plugins/api-accessor'
+  css: [
+    'vuetify/lib/styles/main.sass',
+    '@mdi/font/css/materialdesignicons.min.css',
   ],
-
-  // Auto import components: https://go.nuxtjs.dev/config-components
-  components: true,
-
-  // Modules for dev and build (recommended): https://go.nuxtjs.dev/config-modules
-  buildModules: [
-    // https://go.nuxtjs.dev/typescript
-    '@nuxt/typescript-build',
-    // https://go.nuxtjs.dev/vuetify
-    '@nuxtjs/vuetify',
-    '@nuxtjs/composition-api/module',
-    ['@pinia/nuxt', {}]
-  ],
-
   // Modules: https://go.nuxtjs.dev/config-modules
   modules: [
-    // https://go.nuxtjs.dev/axios
-    '@nuxtjs/axios',
-    // https://go.nuxtjs.dev/pwa
-    // '@nuxtjs/pwa',
-    '@nuxtjs/universal-storage'
+    '@pinia/nuxt',
+    [
+      '@vueuse/nuxt',
+      {
+        ssrHandlers: true,
+      },
+    ],
   ],
-
-  // Axios module configuration: https://go.nuxtjs.dev/config-axios
-  // Axios module configuration (https://go.nuxtjs.dev/config-axios)
-  axios: {
-    proxy: true
-  },
-  proxy: {
-    '/api': {
-      target: process.env.NUXT_ENV_API_URL,
-      pathRewrite: { '^/api/': '/' }
-    }
-  },
-  publicRuntimeConfig: {
-    googleAnalytics: {
-      id: process.env.NUXT_ENV_GOOGLE_ANALYTICS_ID
+  runtimeConfig: {
+    public: {
+      googleApi: {
+        clientId: process.env.NUXT_ENV_GOOGLE_CLIENT_ID,
+        apiKey: process.env.NUXT_ENV_GOOGLE_API_KEY,
+        discoveryDocs: (process.env.NUXT_ENV_GOOGLE_DISCOVERY_DOCS || '').split(
+          ',',
+        ),
+        scopes: process.env.NUXT_ENV_GOOGLE_SCOPES,
+      },
+      googleAnalytics: {
+        id: process.env.NUXT_ENV_GOOGLE_ANALYTICS_ID,
+      },
     },
     baseURL: process.env.BASE_URL,
-    googleApi: {
-      clientId: process.env.NUXT_ENV_GOOGLE_CLIENT_ID,
-      apiKey: process.env.NUXT_ENV_GOOGLE_API_KEY,
-      discoveryDocs: (process.env.NUXT_ENV_GOOGLE_DISCOVERY_DOCS || '').split(
-        ','
-      ),
-      scopes: process.env.NUXT_ENV_GOOGLE_SCOPES
-    }
   },
 
-  storage: {
-    initialState: {
-      myFirstEntry: true
-    },
-    cookie: {
-      prefix: '',
-      options: {
-        path: '/',
-        maxAge: 60 * 60 * 24 * 7 * 4
-      }
-    }
-  } as any,
+  routeRules: {
+    '/api/**': { proxy: `${process.env.NUXT_ENV_API_URL}/**` },
+  },
 
   // Vuetify module configuration: https://go.nuxtjs.dev/config-vuetify
-  vuetify: {
-    optionsPath: '~/config/vuetify.options.js',
-    customVariables: ['~/assets/variables.scss']
+
+  build: {
+    transpile: ['vuetify'],
   },
-  // Build Configuration: https://go.nuxtjs.dev/config-build
-  build: {}
-}
+  hooks: {
+    'vite:extendConfig': (config) => {
+      config.plugins?.push(
+        vuetify({
+          autoImport: true,
+        }),
+      )
+    },
+  },
+})
 
 export default config
