@@ -1,15 +1,58 @@
 <template>
-  <div>
-    <NuxtChild />
-    <v-dialog v-if="firstEntry" v-model="firstEntry" max-width="600" persistent>
-      <InitialSettings :dialog.sync="firstEntry" />
+  <v-container>
+    <v-dialog
+      v-show="firstEntry"
+      v-model="firstEntry"
+      max-width="600"
+      persistent
+    >
+      <InitialSettings />
     </v-dialog>
-  </div>
+    <v-dialog v-model="isNewHourlyLoad" max-width="600">
+      <v-card>
+        <v-card-title class="text-h5">Nueva Carga Horaria</v-card-title>
+        <v-card-text>
+          Se ha encontrado una nueva carga horaria. Actualiza las secciones de
+          los cursos que ya tenias guardadas si deseas los nuevos horarios.
+        </v-card-text>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn
+            color="green-darken-1"
+            variant="text"
+            @click="isNewHourlyLoad = false"
+          >
+            Cerrar
+          </v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <v-dialog v-model="isUpdateHourlyLoad" persistent>
+      <v-card>
+        <v-card-title class="text-h5"
+          >Se ha actualizado la Carga Horaria</v-card-title
+        >
+        <v-card-text>
+          Actualiza las secciones de las materias que ya tenias guardadas si
+          deseas los nuevos horarios.
+        </v-card-text>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn
+            color="green-darken-1"
+            variant="text"
+            @click="isUpdateHourlyLoad = false"
+          >
+            Cerrar
+          </v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <NuxtPage />
+  </v-container>
 </template>
 
 <script lang="ts">
-import { Context } from '@nuxt/types'
-import { useRouter } from '@nuxtjs/composition-api'
 import { storeToRefs } from 'pinia'
 import { defineComponent, watch, onMounted } from 'vue'
 import InitialSettings from '~/components/setting/Initial.vue'
@@ -19,20 +62,31 @@ import { useUserEventsStore } from '~/stores/user-events'
 export default defineComponent({
   name: 'Generator',
   components: {
-    InitialSettings
+    InitialSettings,
   },
-  layout: 'app',
-  setup () {
+  setup() {
+    definePageMeta({
+      layout: 'app',
+    })
+
     const configStore = useUserConfigStore()
     const eventsStore = useUserEventsStore()
     const router = useRouter()
 
-    const { firstEntry } = storeToRefs(configStore)
+    const { firstEntry, isNewHourlyLoad, isUpdateHourlyLoad } =
+      storeToRefs(configStore)
 
-    watch(firstEntry, (newValue, oldValue) => {
+    watch(firstEntry, async (newValue, oldValue) => {
       if (oldValue && !newValue) {
-        router.push('/generator/subjects')
+        await router.push('/generator/subjects')
       }
+    })
+    const store = useUserConfigStore()
+    const { refresh } = useAsyncData(() => {
+      store.fetchFirstEntry()
+      store.fetchFaculty()
+      store.fetchSpeciality()
+      store.fetchHourlyLoad()
     })
 
     onMounted(async () => {
@@ -44,19 +98,15 @@ export default defineComponent({
     })
 
     return {
-      firstEntry
+      firstEntry,
+      isNewHourlyLoad,
+      isUpdateHourlyLoad,
+      refresh,
     }
   },
-  async asyncData ({ $pinia }: Context) {
-    const store = useUserConfigStore($pinia)
-    await store.fetchFirstEntry()
-    await store.fetchFaculty()
-    await store.fetchSpeciality()
-    await store.fetchHourlyLoad()
-  },
   head: {
-    title: 'Generador de Horarios'
-  }
+    title: 'Generador de Horarios',
+  },
 })
 </script>
 
