@@ -2,49 +2,53 @@
   <v-container>
     <h1>Política de privacidad de Horext</h1>
     <div>Ultima revisión: 04 de febrero de 2021</div>
-    <br>
-    <h2>
-      Integración de servicios de terceros
-    </h2>
+    <br />
+    <h2>Integración de servicios de terceros</h2>
     <div class="text">
-      Horext permite integrar y crear comandos para varios servicios de terceros en línea (como se detalla en "servicios de
-      terceros"). Para aprovechar estas funcionalidades, es posible que deba autenticarse, registrarse o iniciar sesión en
-      servicios de terceros a través del Servicio o en los sitios web de sus respectivos proveedores. Cuando habilita la
-      vinculación o inicia sesión en servicios de terceros a través del Servicio, recopilaremos la información relevante
-      necesaria para permitir que el Servicio acceda a ese Servicio de terceros y sus datos y contenido incluidos en ese
-      Servicio de terceros ("Credenciales de inicio de sesión "). Almacenamos sus Credenciales de inicio de sesión el
-      tiempo suficiente para permitir la vinculación al Servicio de terceros.
-      <br>
-      Cuando habilita el Servicio para vincular contenido y datos entre servicios de Terceros, los servicios de Terceros
-      nos proporcionarán acceso a cierta información que puede haber proporcionado a los servicios de Terceros, y
-      usaremos, almacenaremos y divulgaremos tal información de acuerdo con esta Política de privacidad y las reglas que
-      establezca para gobernar el enlace. Además, puede utilizar el Servicio para compartir contenido y Datos personales
-      entre los servicios de terceros que integra con el Servicio. Recuerde que la forma en que los servicios de
-      terceros utilizan, almacenan y divulgan su información se rige por las políticas de dichos servicios de terceros,
-      y Horext no tendrá responsabilidad alguna por las prácticas de privacidad u otras acciones de los servicios de
+      Horext permite integrar y crear comandos para varios servicios de terceros
+      en línea (como se detalla en "servicios de terceros"). Para aprovechar
+      estas funcionalidades, es posible que deba autenticarse, registrarse o
+      iniciar sesión en servicios de terceros a través del Servicio o en los
+      sitios web de sus respectivos proveedores. Cuando habilita la vinculación
+      o inicia sesión en servicios de terceros a través del Servicio,
+      recopilaremos la información relevante necesaria para permitir que el
+      Servicio acceda a ese Servicio de terceros y sus datos y contenido
+      incluidos en ese Servicio de terceros ("Credenciales de inicio de sesión
+      "). Almacenamos sus Credenciales de inicio de sesión el tiempo suficiente
+      para permitir la vinculación al Servicio de terceros.
+      <br />
+      Cuando habilita el Servicio para vincular contenido y datos entre
+      servicios de Terceros, los servicios de Terceros nos proporcionarán acceso
+      a cierta información que puede haber proporcionado a los servicios de
+      Terceros, y usaremos, almacenaremos y divulgaremos tal información de
+      acuerdo con esta Política de privacidad y las reglas que establezca para
+      gobernar el enlace. Además, puede utilizar el Servicio para compartir
+      contenido y Datos personales entre los servicios de terceros que integra
+      con el Servicio. Recuerde que la forma en que los servicios de terceros
+      utilizan, almacenan y divulgan su información se rige por las políticas de
+      dichos servicios de terceros, y Horext no tendrá responsabilidad alguna
+      por las prácticas de privacidad u otras acciones de los servicios de
       terceros. que puede estar habilitado dentro del Servicio.
-      <br>
-      Podemos retener cierta información personal no identificable relacionada con los datos o el contenido vinculado
-      entre los servicios de terceros (por ejemplo, la fecha de envío, la configuración del enlace, los nombres de los
-      servicios de terceros), con el fin de mejorar nuestros servicios y como se describe a continuación en la sección
-      "Datos personales agregados".
+      <br />
+      Podemos retener cierta información personal no identificable relacionada
+      con los datos o el contenido vinculado entre los servicios de terceros
+      (por ejemplo, la fecha de envío, la configuración del enlace, los nombres
+      de los servicios de terceros), con el fin de mejorar nuestros servicios y
+      como se describe a continuación en la sección "Datos personales
+      agregados".
     </div>
-    <br>
-    <h2>
-      Servicios de terceros
-    </h2>
+    <br />
+    <h2>Servicios de terceros</h2>
     <div>
       Actualmente se ocupan los siguientes servicios con terceros
       <ul>
-        <li>Integración con Google Calendar para exportar tus horarios preferidos</li>
+        <li>
+          Integración con Google Calendar para exportar tus horarios preferidos
+        </li>
       </ul>
     </div>
-    <h2>
-      Recopilación de datos
-    </h2>
-    <div>
-      Actualmente no se realiza la recopilación de datos
-    </div>
+    <h2>Recopilación de datos</h2>
+    <div>Actualmente no se realiza la recopilación de datos</div>
   </v-container>
 </template>
 
