@@ -2,22 +2,26 @@
   <v-container>
     <h1>Terminos del servicio de Horext</h1>
     <div>Ultima revisión: 04 de febrero de 2021</div>
-    <br>
-    <h2>
-      Servicios de terceros vinculados
-    </h2>
+    <br />
+    <h2>Servicios de terceros vinculados</h2>
     <div class="text">
-      El Servicio puede contener enlaces e integraciones con sitios web, aplicaciones y servicios de terceros
-      ("Servicios de terceros"), y usted acepta que Horext proporciona enlaces e integraciones con dichos Servicios de
-      terceros únicamente para su conveniencia y no tiene ninguna responsabilidad por el contenido. o disponibilidad de
-      dichos Servicios de terceros, y que Horext no respalda dichos Servicios de terceros (o cualquier producto u otros
-      servicios asociados con ellos). El acceso a los Servicios de terceros vinculados al Servicio es bajo su propio
-      riesgo, y Octatec no es responsable de la precisión o confiabilidad de la información, los datos, las opiniones,
-      los consejos o las declaraciones realizadas sobre dichos Servicios de terceros. Su uso de dichos Servicios de
-      terceros estará sujeto a los términos aplicables a cada uno de dichos Servicios de terceros. No puede publicar un
-      enlace que dirija a los usuarios a cualquier contenido o información que constituya una violación de estos
-      Términos de servicio o cualquier término aplicable a dichos Servicios de terceros
-      </div>
+      El Servicio puede contener enlaces e integraciones con sitios web,
+      aplicaciones y servicios de terceros ("Servicios de terceros"), y usted
+      acepta que Horext proporciona enlaces e integraciones con dichos Servicios
+      de terceros únicamente para su conveniencia y no tiene ninguna
+      responsabilidad por el contenido. o disponibilidad de dichos Servicios de
+      terceros, y que Horext no respalda dichos Servicios de terceros (o
+      cualquier producto u otros servicios asociados con ellos). El acceso a los
+      Servicios de terceros vinculados al Servicio es bajo su propio riesgo, y
+      Octatec no es responsable de la precisión o confiabilidad de la
+      información, los datos, las opiniones, los consejos o las declaraciones
+      realizadas sobre dichos Servicios de terceros. Su uso de dichos Servicios
+      de terceros estará sujeto a los términos aplicables a cada uno de dichos
+      Servicios de terceros. No puede publicar un enlace que dirija a los
+      usuarios a cualquier contenido o información que constituya una violación
+      de estos Términos de servicio o cualquier término aplicable a dichos
+      Servicios de terceros
+    </div>
   </v-container>
 </template>
 
