@@ -1,20 +1,20 @@
-import { AxiosInstance } from 'axios'
+import type { $Fetch } from 'nitropack'
 const PATH_SUBJECTS = 'classSessions'
-export default ($axios: AxiosInstance) => ({
-  findScheduleId (schedule:number) {
-    return $axios.get(PATH_SUBJECTS,
-      {
-        params: {
-          schedule
-        }
-      })
+export default ($axios: $Fetch) => ({
+  findScheduleId(schedule: number) {
+    return $axios(PATH_SUBJECTS, {
+      method: 'GET',
+      params: {
+        schedule,
+      },
+    })
   },
-  findScheduleIds (schedulesIds:Array<number>) {
-    return $axios.get(PATH_SUBJECTS,
-      {
-        params: {
-          schedules: schedulesIds.join(',')
-        }
-      })
-  }
+  findScheduleIds(schedulesIds: Array<number>) {
+    return $axios(PATH_SUBJECTS, {
+      method: 'GET',
+      params: {
+        schedules: schedulesIds.join(','),
+      },
+    })
+  },
 })
