@@ -1,9 +1,9 @@
-
-import { AxiosInstance } from 'axios'
+import type { $Fetch } from 'nitropack'
+import type { IOrganization } from '~/interfaces/organization'
 
 const PATH_SUBJECTS = 'faculties'
-export default ($axios: AxiosInstance) => ({
-  getAll () {
-    return $axios.get(PATH_SUBJECTS)
-  }
+export default ($axios: $Fetch) => ({
+  getAll() {
+    return $axios<IOrganization[]>(PATH_SUBJECTS)
+  },
 })
