@@ -1,12 +1,12 @@
-
-import { AxiosInstance } from 'axios'
+import type { $Fetch } from 'nitropack'
+import type { IOrganization } from '~/interfaces/organization'
 const PATH_SUBJECTS = 'specialities'
-export default ($axios: AxiosInstance) => ({
-  getAllByFaculty (facultyId: any) {
-    return $axios.get(PATH_SUBJECTS, {
+export default ($axios: $Fetch) => ({
+  getAllByFaculty(facultyId: any) {
+    return $axios<IOrganization[]>(PATH_SUBJECTS, {
       params: {
-        faculty: facultyId
-      }
+        faculty: facultyId,
+      },
     })
-  }
+  },
 })
