@@ -1,5 +1,4 @@
-
-import { AxiosInstance } from 'axios'
+import type { $Fetch } from 'nitropack'
 import CourseRepository from '~/repositories/CourseRepository'
 import ScheduleSubjectRepository from '~/repositories/ScheduleSubjectRepository'
 import ClassSessionRepository from '~/repositories/ClassSessionRepository'
@@ -7,11 +6,11 @@ import FacultyRepository from '~/repositories/FacultyRepository'
 import HourlyLoadRepository from '~/repositories/HourlyLoadRepository'
 import SpecialityRepository from '~/repositories/SpecialityRepository'
 
-export default ($axios: AxiosInstance) => ({
+export default ($axios: $Fetch) => ({
   course: CourseRepository($axios),
   classSessions: ClassSessionRepository($axios),
   speciality: SpecialityRepository($axios),
   faculty: FacultyRepository($axios),
   hourlyLoad: HourlyLoadRepository($axios),
-  scheduleSubject: ScheduleSubjectRepository($axios)
+  scheduleSubject: ScheduleSubjectRepository($axios),
 })
