@@ -1,9 +1,11 @@
 import { DateTime } from 'luxon'
+import type { ISchedule } from '~/interfaces/schedule'
+import type { ISelectedSubject, ISubjectSchedule } from '~/interfaces/subject'
 import Event from '~/model/Event'
 
 const isIntersects = (
   eventTarget: Event,
-  eventSource: { start: string; end: string }
+  eventSource: { start: string; end: string },
 ): boolean =>
   !(
     eventTarget.end <= eventSource.start || eventSource.end <= eventTarget.start
@@ -22,20 +24,27 @@ export const convertToDate = (day: string | number, startTime: string) => {
   return weekdayToDatetime(<number>day, startTime)
 }
 
-export function getSchedules (
-  subjects: Array<any>,
-  myEvents: Array<any>,
-  options: any = {
+export type ScheduleOptions = {
+  credits?: number
+  crossingSubjects: number
+  crossEvent?: boolean
+  crossPractices?: boolean
+}
+
+export function getSchedules(
+  subjects: Array<ISelectedSubject>,
+  myEvents: Array<Event>,
+  options: ScheduleOptions = {
     credits: 100,
     crossingSubjects: 0,
     crossEvent: true,
-    crossPractices: false
-  }
-): { occurrences: any[]; schedules: any[]; combinations: any[] } {
+    crossPractices: false,
+  },
+): { occurrences: any[]; schedules: any[]; combinations: ISchedule[] } {
   const occurrences = []
   const maxQuantity = subjects.length
   const indexSchedules = Array(maxQuantity).fill(0)
-  const schedules: Array<any> = []
+  const schedules: Array<ISchedule> = []
 
   const increment = (i: number) => {
     if (i >= 0 && indexSchedules[i] === subjects[i].schedules.length - 1) {
@@ -50,19 +59,19 @@ export function getSchedules (
     return previousValue * currentValue.schedules.length
   }, 1)
   const crossings = Array(combinations).fill(0)
-  for (let i = combinations; i--;) {
+  for (let i = combinations; i--; ) {
     const combination: Array<any> = []
     for (let j = 0; j < indexSchedules.length; j++) {
       const subject = subjects[j]
       const schedule = subjects[j].schedules[indexSchedules[j]]
       combination.push({
         ...schedule,
-        subject
+        subject,
       })
     }
     const currentSchedule = combination.map((c, index) => ({
       ...c,
-      events: scheduleToEvent(c, colors[index])
+      events: scheduleToEvent(c, colors[index]),
     }))
     // calculating crossing
     let crossingCombination = 0
@@ -79,7 +88,7 @@ export function getSchedules (
             occurrences.push({
               type: 'Cruce de ' + event.title + ' - ' + item.title,
               elementA: event,
-              elementB: item
+              elementB: item,
             })
             // if have available crossings
             if (
@@ -113,14 +122,14 @@ export function getSchedules (
     if (crossingCombination <= options.crossingSubjects && useCombination) {
       crossings[i] = crossingCombination
       schedules.push({
-        id: combination.map(c => c.scheduleSubject.id).join(','),
-        scheduleSubjectIds: combination.map(c => c.scheduleSubject.id),
+        id: combination.map((c) => c.scheduleSubject.id).join(','),
+        scheduleSubjectIds: combination.map((c) => c.scheduleSubject.id),
         schedule: combination,
         crossings: crossingCombination,
         events: combination
           .map((c, index) => scheduleToEvent(c, colors[index]))
           .flat()
-          .concat(myEvents)
+          .concat(myEvents),
       })
     }
 
@@ -130,13 +139,13 @@ export function getSchedules (
   return {
     schedules: [],
     combinations: schedules,
-    occurrences
+    occurrences,
   }
 }
 
-function scheduleToEvent (
-  schedule: any,
-  color: string = 'primary'
+function scheduleToEvent(
+  schedule: ISubjectSchedule,
+  color = 'primary',
 ): Array<Event> {
   const events: Array<Event> = []
   const sessions = schedule?.sessions || []
@@ -152,7 +161,7 @@ function scheduleToEvent (
       sessions[i]?.classroom?.code,
       color,
       'COURSE',
-      sessions[i].type.code
+      sessions[i].type.code,
     )
     events.push(event)
   }
@@ -163,38 +172,38 @@ const weekDays = [
   {
     index: 0,
     value: 'Domingo',
-    startDate: '2020-11-15'
+    startDate: '2020-11-15',
   },
   {
     index: 1,
     value: 'Lunes',
-    startDate: '2020-11-09'
+    startDate: '2020-11-09',
   },
   {
     index: 2,
     value: 'Martes',
-    startDate: '2020-11-10'
+    startDate: '2020-11-10',
   },
   {
     index: 3,
     value: 'Miercoles',
-    startDate: '2020-11-11'
+    startDate: '2020-11-11',
   },
   {
     index: 4,
     value: 'Jueves',
-    startDate: '2020-11-12'
+    startDate: '2020-11-12',
   },
   {
     index: 5,
     value: 'Viernes',
-    startDate: '2020-11-13'
+    startDate: '2020-11-13',
   },
   {
     index: 6,
     value: 'Sábado',
-    startDate: '2020-11-14'
-  }
+    startDate: '2020-11-14',
+  },
 ]
 export { weekDays }
 
@@ -216,7 +225,7 @@ export const colors = [
   'blue darken-3',
   'green lighten-3',
   'orange lighten-3',
-  'blue lighten-3'
+  'blue lighten-3',
 ]
 
 export const weekdays = [
@@ -226,5 +235,5 @@ export const weekdays = [
   'Miercoles',
   'Jueves',
   'Viernes',
-  'Sábado'
+  'Sábado',
 ]
