@@ -1,3 +1,3 @@
-export default function timeout (ms: number | undefined) {
-  return new Promise(resolve => setTimeout(resolve, ms))
+export default function timeout(ms: number | undefined) {
+  return new Promise((resolve) => setTimeout(resolve, ms))
 }
