@@ -1,8 +1,5 @@
 // Utilities
-import {
-  camelCase,
-  upperFirst,
-} from 'lodash'
+import { camelCase, upperFirst } from 'lodash'
 
 export default function (sections = []) {
   return {
